+++ src/Web/Models/InventoryDtos.cs
@@ -21,9 +21,9 @@ public class StockMovementDto
     public int WarehouseId { get; set; }
     public string WarehouseName { get; set; } = string.Empty;
     public string MovementType { get; set; } = "ajuste";
-    public int DeltaQuantity { get; set; }
-    public int StockBefore { get; set; }
-    public int StockAfter { get; set; }
+    public decimal DeltaQuantity { get; set; }
+    public decimal StockBefore { get; set; }
+    public decimal StockAfter { get; set; }
     public string? Reason { get; set; }
     public string? Notes { get; set; }
     public string? OperatorName { get; set; }
@@ -35,7 +35,7 @@ public class AdjustStockRequest
     public int ProductId { get; set; }
     public int WarehouseId { get; set; }
     public string MovementType { get; set; } = "ajuste";
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public string? Reason { get; set; }
     public string? Notes { get; set; }
     public string? OperatorName { get; set; }
